pyproject.toml
[project]
name = "pickle-bot"
version = "0.1.0"
description = "Personal AI assistant with pluggable skills"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "litellm>=1.0.0",
  "typer>=0.12.0",
  "textual>=0.21.0",
  "pydantic>=2.0.0",
  "pyyaml>=6.0",
  "python-dotenv>=1.0.0",
  "rich>=13.0.0",
]

[project.scripts]
picklebot = "picklebot.cli.main:app"

[tool.uv.sources]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/picklebot"]
